requirements.txt
pytz
bcrypt==3.1.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.2
Flask-Session==0.3.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pycparser==2.20
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
